Switch gui.py to be main.py
